Move visual tests run to visual.spec
