+++ lib/KIcon/__tests__/visual.spec.js
