+++ lib/KTextbox/__tests__/KTextbox.spec.js
@@ -1,9 +1,4 @@
 import { shallowMount, mount } from '@vue/test-utils';
-import {
-  click,
-  renderComponentForVisualTest,
-  takeSnapshot,
-} from '../../../jest.conf/visual.testUtils';
 import KTextbox from '../index';
 
 describe('KTextbox component', () => {
@@ -223,17 +218,5 @@ describe('KTextbox component', () => {
       await wrapper.vm.$nextTick();
       expect(wrapper.find('input').element.value).toBe('');
     });
-  });
-});
-
-describe.visual('KTextbox visual tests', () => {
-  const snapshotOptions = { widths: [800], minHeight: 512 };
-  it('renders', async () => {
-    await renderComponentForVisualTest('KTextboxVisualTest');
-
-    // Prepares active textbox test case
-    await click('#active-textbox input');
-
-    await takeSnapshot('KTextbox visual tests', snapshotOptions);
   });
 });
